requirements.txt
# MCP Framework
fastmcp>=0.1.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Web Server
uvicorn>=0.24.0
gunicorn>=21.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.2